fixed image-date and added date taken option
When running an import with history (more than one day) all images were
put on the current day; now all images are put on their respective days.
Added a configuration option (flickr_datetype) that can be set to
'taken' it then fetches the image taken date instead of the upload-date.