No time zone support needed.
What I thought was a bug in the trakt feed was a misunderstanding.  On a check-in they put the time that a person would finish watching a show in the feed:  current time + length of show.  If you want the feed to reflect the time you started the show, mark it as "previously watched".