Set Omnifocus to read from 00:00:00.
The Omnifocus plugin relies on a setDate method to build a date on which
to look for completed tasks. However, setDate doesn't force a time
portion of the date and this causes AppleScript to default to the
current time of day. Thus, unless a task has been marked as completed in
the future, the plugin couldn't never return results for the current day. The
plugin also couldn't return entries from the first day of any multi-day
runs that occur before the present hh:mm:ss part of that day.

If I complete a task in Omnifocus at 13:01 and run slogger two minutes
later the effect is to ask Omnifocus, "what have I completed today,
since 13:03?", which is nothing. If I run slogger -t1 the next day at
15:00, I'm asking, "what did I complete yesterday, after 15:00?"

This is fixed by explicity setting the time portion of the cut-off date
to the very beginning of they day, 00:00:00