Update fitgem to support oauth2 (#420)
* Update fitbit plugin to support oauth2

This changes the authentication method of this gem from oauth1 to
oauth2. This needs the git version of the fitbit gem. The easiest way to
get this is to add the following line to the Slogger Gemfile and run
bundle update:

gem 'fitgem', :git => 'git://github.com/whazzmaster/fitgem'

* Update documentation for fitgem

Updated the documentation for fitgem. This fitgem is no longer being
supported and version 1.0.0 may not end up on rubygems.org. Right now it
needs to be installed from https://github.com/whazzmaster/fitgem.

* Update fitgem installation instructions